+++ examples/rtk_example.py
@@ -18,7 +18,7 @@ to an RTK-compatible u-blox GNSS receiver (e.g. ZED-F9P)
 connected to a local serial port (USB or UART1).
 
 GNSSNTRIPClient receives RTCM3 or SPARTN data from the NTRIP
-caster and outputs it to a message queue. An example
+caster and outputs it to a message queue. A basic
 GNSSSkeletonApp class reads data from this queue and sends
 it to the receiver, while reading and parsing data from the
 receiver and printing it to the terminal.
@@ -26,6 +26,8 @@ receiver and printing it to the terminal.
 GNSSNtripClient optionally sends NMEA GGA position sentences
 to the caster at a prescribed interval, using either fixed
 reference coordinates or live coordinates from the receiver.
+For NTRIP 2.0 protocol, the first GGA sentence is embedded
+in the HTTP GET request header.
 
 NB: Some NTRIP casters may stop sending RTK data after a while
 if they're not receiving legitimate NMEA GGA position updates
@@ -40,15 +42,25 @@ Created on 5 Jun 2022
 
 # pylint: disable=invalid-name
 
+from logging import getLogger
+from queue import Empty, Queue
 from sys import argv
-from queue import Queue, Empty
 from threading import Event
 from time import sleep
 
-from pygnssutils import VERBOSITY_LOW, GNSSNTRIPClient
 from gnssapp import GNSSSkeletonApp
 
+from pygnssutils import (
+    VERBOSITY_DEBUG,
+    VERBOSITY_HIGH,
+    VERBOSITY_MEDIUM,
+    GNSSNTRIPClient,
+    set_logging,
+)
+
 CONNECTED = 1
+
+logger = getLogger("pygnssutils")
 
 
 def main(**kwargs):
@@ -57,7 +69,7 @@ def main(**kwargs):
     """
 
     # GNSS receiver serial port parameters - AMEND AS REQUIRED:
-    SERIAL_PORT = "/dev/ttyACM0"
+    SERIAL_PORT = "/dev/ttyACM0"  # use "UBXSIMULATOR" to use dummy UBX serial stream
     BAUDRATE = 38400
     TIMEOUT = 10
 
@@ -86,10 +98,12 @@ def main(**kwargs):
     recv_queue = Queue()  # data from receiver placed on this queue
     send_queue = Queue()  # data to receiver placed on this queue
     stop_event = Event()
-    verbosity = 0  # 0 - no output, 1 - print identities, 2 - print full message
+
+    set_logging(logger, VERBOSITY_HIGH)
+    mylogger = getLogger("pygnssutils.rtk_example")
 
     try:
-        print(f"Starting GNSS reader/writer on {SERIAL_PORT} @ {BAUDRATE}...\n")
+        mylogger.info(f"Starting GNSS reader/writer on {SERIAL_PORT} @ {BAUDRATE}...\n")
         with GNSSSkeletonApp(
             SERIAL_PORT,
             BAUDRATE,
@@ -97,15 +111,14 @@ def main(**kwargs):
             stopevent=stop_event,
             recvqueue=recv_queue,
             sendqueue=send_queue,
-            verbosity=verbosity,
             enableubx=True,
             showstatus=True,
         ) as gna:
             gna.run()
             sleep(2)  # wait for receiver to output at least 1 navigation solution
 
-            print(f"Starting NTRIP client on {NTRIP_SERVER}:{NTRIP_PORT}...\n")
-            with GNSSNTRIPClient(gna, verbosity=VERBOSITY_LOW) as gnc:
+            mylogger.info(f"Starting NTRIP client on {NTRIP_SERVER}:{NTRIP_PORT}...\n")
+            with GNSSNTRIPClient(gna) as gnc:
                 streaming = gnc.run(
                     ipprot=IPPROT,
                     server=NTRIP_SERVER,
@@ -134,10 +147,6 @@ def main(**kwargs):
                         try:
                             while not recv_queue.empty():
                                 (_, parsed_data) = recv_queue.get(False)
-                                if verbosity == 1:
-                                    print(f"GNSS>> {parsed_data.identity}")
-                                elif verbosity == 2:
-                                    print(parsed_data)
                                 recv_queue.task_done()
                         except Empty:
                             pass
@@ -146,7 +155,7 @@ def main(**kwargs):
 
     except KeyboardInterrupt:
         stop_event.set()
-        print("Terminated by user")
+        mylogger.info("Terminated by user")
 
 
 if __name__ == "__main__":